Host tests for RX hardware-bound modules: button, display, gps, sd_card
New test suites (fakes simulate the hardware surface; module sources are
#included for static access):
- test_button.c: EXTI+GPIO simulation - debounce time gate, release-bounce
  state guard, short-press (fires at release) vs long-press (fires while
  held), button 2 fire-on-press, stuck-button retrigger block.
- test_display.c: capturing I2C fake - init sequence, framebuffer pixel/
  line/rect/circle/char correctness, right-edge text clipping, update
  chunking (6 cmds + 64x17-byte chunks), and the headless-mode regression
  (I2C init failure => zero bus traffic, receiver keeps running).
- test_gps.c: UART fake driving the real ISR callback - NMEA sentence
  assembly, framing-error rejection, buffer-overflow reset, mid-sentence
  restart, IRQ re-arm paths, debug rotation. Links real gps_parser.c.
- test_sd_card.c: REAL littlefs over a RAM block device - mount-retry-
  then-format, bd error codes 91/92/99/13, lazy log-file creation with
  L000N sequencing, all CSV row formats (incl. fused nav rows), beacon +
  compass-cal persistence with corruption rejection, format, self-test.

Firmware fixes found by/for these tests:
- gps.c: UART init failure now returns GPS_ERROR (degrade path in main)
  instead of calling Error_Handler, which since the fault-handler change
  would have reset the receiver over an optional peripheral.
- sd_card.c/gps.c: %lu format args cast to unsigned long (host-portable;
  was UB on LP64 hosts, benign on ARM).

Stub HAL extended with EXTI/I2C/UART surfaces. RX suite now 8 binaries,
1041 checks, all green. ARM build unaffected.

+++ receiver/firmware/src/sd_card.c
@@ -344,7 +344,8 @@ void SD_Card_GetTimestamp(char *timestamp_str, uint32_t max_len)
 {
     if (timestamp_str == NULL) return;
     uint32_t tick = HAL_GetTick();
-    snprintf(timestamp_str, max_len, "%lu.%03lu", tick / 1000, tick % 1000);
+    snprintf(timestamp_str, max_len, "%lu.%03lu",
+             (unsigned long)(tick / 1000), (unsigned long)(tick % 1000));
 }
 
 SD_Card_Status SD_Card_LogGPS(GPS_Data *gps_data, Log_Entry_Type log_type)
